+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/DemonHunter/JaceDarkweaver.cs
@@ -11,7 +11,8 @@ public class JaceDarkweaver: ICardWithRelatedCards
 	public bool ShouldShowForOpponent(Player opponent)
 	{
 		var card = Database.GetCardFromId(GetCardId());
-		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentFormat, opponent.OriginalClass) && GetRelatedCards(opponent).Count > 3;
+		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentGameType, Core.Game.CurrentFormatType, opponent.OriginalClass)
+		       && GetRelatedCards(opponent).Count > 3;
 	}
 	public List<Card?> GetRelatedCards(Player player) =>
 		player.SpellsPlayedCards