+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Druid/HydrationStation.cs
@@ -11,7 +11,7 @@ public class HydrationStation: ICardWithRelatedCards
 	public bool ShouldShowForOpponent(Player opponent)
 	{
 		var card = Database.GetCardFromId(GetCardId());
-		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentFormat, opponent.OriginalClass) && GetRelatedCards(opponent).Count > 1;
+		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentGameType, Core.Game.CurrentFormatType , opponent.OriginalClass) && GetRelatedCards(opponent).Count > 1;
 	}
 
 	public List<Card?> GetRelatedCards(Player player) =>