+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Priest/RestInPeace.cs
@@ -10,7 +10,7 @@ public class RestInPeace: ICardWithRelatedCards
 	public bool ShouldShowForOpponent(Player opponent)
 	{
 		var card = Database.GetCardFromId(GetCardId());
-		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentFormat, opponent.OriginalClass) && GetRelatedCards(opponent).Count > 0;
+		return CardUtils.MayCardBeRelevant(card, Core.Game.CurrentGameType, Core.Game.CurrentFormatType , opponent.OriginalClass) && GetRelatedCards(opponent).Count > 0;
 	}
 	private string OpponentHero => Core.Game.CurrentGameStats?.OpponentHeroCardId ?? "";